backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "google-cloud-storage>=3.8.0",
    "gunicorn>=24.1.1",
    "numpy>=2.4.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "tensorflow>=2.20.0",
    "tensorflow-hub>=0.16.1",
]
